OneDrive/Desktop/trae_trial/tomato_blight_app/backend/requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
joblib>=1.3.2
h5py>=3.10.0